Start adding linearity checks.
We'll get better error messages if we do it in the type inference pass (and we
may eventually want to do some linearity inference too) but that makes it more
complicated so it's in Type.hs for now. For testing, I added a `:tycheck`
command to check the type (including linearity) after the inference pass.