Add a `Data` typeclass for anything that's not a function. Fixes #27.
One complication is that the interpreter uses the table constructor with
function arguments internally, which violates this new check. The interpreter
only needs to query types, not check them, so we'll be fine if we reintroduce
the "query without checking" mode of the type checker. I'll do that in a
separate commit.