Give more precise linearity rules to scan.
We can now give linear types to `sum`, `mean` etc. It turned out to be a bit
trickier than expected. Our previous formulation of scan operated on a table of
functions, which concealed linearity. The problem is that linear variables can
enter the function body by two different routes (lexical scope and formal
parameters) but the typing rules don't know that they have the same source. It's
essentially the difference between these two:

   y = x; x + y       -- this is considered linear in x
   (lam y. x + y) x   -- this is not. The function by itself isn't linear in x.