Add some light linearity inference.
Type variables representing multiplicities become part of constraints which are
solved in the usual way. Multiplicity tags (`Lin`/`NonLin`) are now just extra
constructors of `Type`. It's a shame to not distinguish multiplicities from
types at the (Haskell) type level, but they have so much in common (variables,
substitution, abstraction, instantiation) that it seemed worse to duplicate all
that machinery.

Our insistence that un-annotated lets be monomorphic caused a bit of trouble
because it's easy to end up with an ambiguous multiplicity. I put in a
defaulting rule that makes ambiguous linearity variables nonlinear. We might
want to revise this.