Rework type inference solver for better error reporting.
Our `solveLocalMonomorphic` was obscuring the source locations of errors by
solving constraints out of order. It was made worse by the extra fresh variables
introduced by recent changes. Now we just unify as we go, similar to an earlier
design. Importantly, we still treat the solver state as a monoid, which lets us
isolate the fresh variables and substitution associated with solving a local
subexpression in order to check leaks and enforce monomorphism.