Add a linear reader effect and switch from linear to affine types.
Our linear reader effect, `{LinReader r} a`, is isomorphic to `r --o a`. It
guarantees that the linear reading environment is used at most once.

Our effect system requires that we can freely add extra effects to a function's
type. For most effects, this is fine. The function just won't exercise them,
e.g. ignoring its reader environment. But a strictly linear reader effect can't
be ignored since it makes the function promise to consume the reader
environment.

The problem goes away if we make the type system affine instead of linear. We
can still transpose an affine function. We just ignore the constant offset and
transpose the linear part (thanks Chris Leary for pointing this out!). For now
we'll keep the lollipop symbol and keep using the word "linear" when we mean
"affine" (it's common informal usage anyway).

Co-authored-by: Matthew Johnson <mattjj@google.com>