Add type-level labels to distinguish effects from each other.
Types of kind `Label` can be either literal names or type variables. Functions
like `runReader` and `ask` are polymorphic over their effects' labels, but you
have to use explicit type application to pass them a specific label. There's no
label inference.

There's a uniqueness requirement that's not currently enforced. In-scope label
variables are assumed to correspond to distinct labels, which means we need to
prohibit applying a function to the same label twice. This is a sort of
linearity requirement on label variables. It should be simpler than linearity
checking on term variables because label variables are type variables, which are
much more restricted.