Defunctionalize through effect dischargers, `runReader` etc.
This lets us have effectful computations that produce functions, which is
especially important for linearization, which generates types like
`(b, a --o b)`.

We're using the "ionization" technique we got rid of in 002b32cb1c. When we have
a function that returns a lambda, we instead return just the local environment
variables and reconstruct the lambda at the call site.