Replace type-level effect labels with term-level ones.
Discharging effects now looks a bit like Haskell's `runST`, except that we're
using dependent types instead of higher-rank polymorphism:

    runReader :: r -> (Pi ref::(Ref r) -> {Reader ref |eff} a) -> {|eff} a

`ref` is a term variable bound by `Pi`, and it appears in the effect type of the
function. Reference variables are used like references in an imperative language
(e.g. `x = load ref`) but they're constrained by a not-yet-implemented linearity
system (yes, another one..) to prevent aliasing.

Type inference works in simple cases but it's still very rough. Pi types
introduce many of the same issues as higher rank polymorphism (e.g. those
discussed in "Practical type inference for arbitrary-rank types") and it'll take
more work to address them. One difference in our favor is that application of pi
types is explicit in the source language whereas with higher-rank polymorphism
application of "forall" types has to be inferred, which creates tricky issues
around subsumption, e.g. coercing `forall a. a -> a` to `forall b. [b] -> [b]`.

We brought back the dreaded de Bruijn indices for pi-bound type variables (we
only lasted 2 weeks without them) because it's so much simpler not to have to
create Skolem variables to unify pi types.

We punted on linearizing effects in this new formulation. We need to think
harder about scoping and name leaking.