Linearization for `for`.
This is a start but it's still very buggy. What's hard is that the tangent
computation becomes a separate `for`, not under the scope of the primal `for`,
so we need to explicitly capture the necessary intermediate values from the
primal `for` and forward them to the tangent part. So `for i. ...` becomes
something like:

   -- primal part
   (ans, residuals) = unzip $ for  i. x1 = ...
                                      x2 = ...
                                      ...
                                      (xAns, (x1, x2, ...))
   -- tangent part
   for i. (x1, x2, ...) = residuals.i
          t1 = ...
          t2 = ...
          ...
          tAns

Currently, this requires a lot of manual management of scopes and substitutions,
and it's clearly not quite right yet.