Linearization rules for reader and writer effects.
As with `for`, the hard part is saving the residuals. Previously we built a
temporary version of the linear expression to query its free variables. This
becomes harder with effects (for uninteresting reasons) so now we just track the
variables needed explicitly.

Also switch the prelude implementation of `sum` to use the writer effect.