Indentation-sensitive parsing of blocks.
Now that we have effects, it makes sense to be able to sequence effectful
expressions without dummy binders: `<expr>; <expr>` instead of
`_ = <expr>; _ = <expr>`.

That requires a new way to determine where a sequence of expressions ends,
since we can't use the decls-followed-by-single-expr pattern. That pattern
was also a bit fragile.

This commit adds indentation rules similar to Python's. The body of `lam`,
`for`, `:p` etc. can be an ordinary expression on the same line, or a "block" of
expressions or let bindings sharing an indentation level.

For example, we can now write:

    runWriter lam w.
      for i::2.
        tell w 1.0
      tell w 1.0

The indentation of the second `tell` determines whether it's part of the loop
body or part of the lambda.