Clarify substitution discipline in Imp lowering by limiting it to a few places.
Adding `Inject` exposed a substitution bug in Imp lowering (missing a
substitution of the type annotation on a `For` binder, as it turned out).

It was hard to identify the bug because every function in the `ImpM` monad had
access to the substitution environment and it wasn't clear whose job it was to
actually apply it!

The solution, which we might want to apply elsewhere, was to take the
substitution environment out of the monad and make it an explicit argument to
the few functions that actually need it. The "triple-functor" structure of our
`PrimOp` ADT turned out to be very convenient. The triply-parameterized
traversal lets us apply the substitution generically to the type-leaves and
atom-leaves (both cheap) while leaving the lambda-leaves as a
`(LamExpr, SubstEnv)` pair, a sort of unforced lazy substitution thunk, which
avoids the cost of repeatedly substituting under a binder. It decluttered the
lowering rules for each primitive operation by quite a bit.