Rework top-level driver (again).
As usual, the goal is to make it easier to add passes, this time motivated by
the JAX/XLA backend.

The main change is to move away from our "module" system where lower-level IRs
were wrapped up with data structures describing how to restructure the resulting
flat arrays into the more complex types of the source language. Since that
mapping is completely determined by the type, we can let the top level driver
work it out, relieving the lower-level passes of the burden of maintaining the
restructuring information.