Fix freeVars to ignore DeBruijn indices
DeBruijn names cannot be free by definition. It would be a bit nicer to
discard them at the binders, but this approach is significantly simpler.

This also incldues a minor cleanup in how `Var`s are handled in
freeVars. The previous instance was binder-specific, so I've removed it
and replaced it with explicit functions that can be used depending on
the semantics of the call site.