Implement RPC-over-pipe for the JAX backend.
This lets you spin up an RPC server subprocess that exposes a collection of
functions. We communicate with the server over a pipe using line-delimited JSON.

This time we avoid clobbering stdin/stdout so we can drop into a debugger on the
Python side. The Python server occasionally raises "BdbQuit" so something's not
quite right.