Minor cleanup and reorg of Type.hs
As I was reading the file to better understand the type checking
machinery, I thought it might be a nice occasion to do a small cleanup
too. Most of it is just shuffling existing code around, to better group
it with relevant functions. I also got rid of all the `check*` functions
that were not part of `HasType` typeclass, because I thought it's a bit
weird that they follow the same naming, while each has its own slightly
different meaning.