Start building a new front-end language, `UExpr`, based around dependent types.
We recently added dependent types for ranges and effects. If we lean on them
more heavily, we can get rid of several specialized type system features. We can
use pi types instead of forall types, Idris-style implicit arguments in place of
implicit type application, and ordinary let bindings in place of type aliases.
We can also de-duplicate implementation machinery that's replicated at the term
level and the type level (e.g. getType/getKind).

Adam Paszke has been patiently suggesting this approach for a while, and I'm
finally starting to understand what a good idea it is.

It would be a big change to do atomically in-place, so we're taking the John
Carmack "parallel implementation" approach, implementing the `UExpr` language
from scratch, alongside the existing `FExpr`. A command-line flag `--frontend
uexpr` swaps out the parser and type inference pass to use the new front-end
(everything after normalization is the same). Once it's feature-complete, we'll
have the satisfaction of deleting the `FExpr` implementation.