Build out UExpr parser.
This time I'm trying to be more careful about making sure that lexeme parsers
don't overlap with each other, to avoided misparses like `:=` as `:` and `=`.

We're also allowing the definitions (but not the fixity and associativity) of
infix ops like `+` to be provided in the prelude instead of built into the
parser, so the parser itself can be simpler.