Small readjustments to arrow parameterization.
The main change is to put effect annotations under the `PlainArrow` case of the
`Arrow` parameter. Now table arrows are syntactically prevented from carrying
effects, which gives us a little more safety.

Also remove the arrow parameter to application. This means that function
application and table indexing are only distinguishable by querying the type of
the lhs. It makes the AST less redundant but it makes pattern-matching on table
indexing more awkward. We may end up reverting it.