Start implementing Idris-like typeclass dictionary synthesis.
We introduce a new implicit arrow, `?=>` for implicit arguments that are
typeclass dictionaries. Whereas `?->` arguments are filled in using unification
(usually they're types) these `?=>` dictionary arguments are synthesized from
in-scope definitions. It's an error if there's more than one solution.

There's currently an interesting failure when we try to synthesize a dictionary
for `Eq (Int & Int)`. There are two solutions for the second component's
dictionary: it could use the top-level instance, or the newly-synthesized
dictionary for the first component! To solve this, I'm going to try adding
annotations to let bindings so that we can limit the available candidates.

Also add a generic parameterized traversal over the core IR for passes like this
dictionary synthesis pass that only care about one or two constructors.