Add validity checks for syntactic variations of our core IR.
These are supposed to capture the sort of simple restrictions that we could
easily enforce statically if we used a different Haskell data type for each
variant, at the cost of more boilerplate.

We hope this makes it easier to reason about the valid inputs to different
passes, especially the Imp pass, which consumes the restricted Simp IR.

I also refactored the `Module` data type a bit, to separate unevaluated decls,
which are subject the restrictions of the IR variants, from evaluated ones,
which are subject to different restrictions, like only binding global names.