Cache prelude evaluation.
As we're putting more in userspace, it's taking longer to evaluate the prelude.
It was almost a full second between `dex repl` and seeing the prompt, and it's
only going to get worse.

This change adds a caching layer using the `store` library. We save the
post-prelude-eval state in `prelude.cache`. We consider the cache fresh if its
modification time is newer than the prelude's, and the Dex compiler we're using
is the same binary as the one used to create the cache (based on a compile-time
timestamp).