Handle multi-constructor types and start handling multi-case case expressions.
As with our built-in `Either`, we (inefficiently) represent sums as a product of
all the constructor payloads, with a run-time integer tag to indicate which one
is valid.

We currently evaluate all branches of a case expression with the intention of
picking the correct result using `select`. But rather than figure out the
selection logic for multi-case cases I'm just going to add a proper switch
statement to Imp in a follow-up change.