Pattern-matching for variants.
Since we are now matching both DataCons and Variants, we can't use the
name of the con to determine the order. Instead, we order alternatives
based on either the index of the con, or the index of the type in the
label map for the variant.

Since you can't tell the type of a variant from its syntax alone,
there's a bit of extra zonking to make sure that we know the type of
the expression to unpack before we check the patterns we are unpacking.

There are a few times where I compute indices into variant types by
enumerating their contents. This will probably be fairly negligible
performance-wise but we could try computing that in a more efficient
way if we need to for some reason.