Revert "Link the Dex runtime statically to all LLVM programs we compile" (#199)
For some reason, LLVM fails to emit proper PIC code for the DexRT code,
because it loads the offset of a statically allocated constant from the
GOT, but then forgets to offset it by the base offset for the whole
binary (which it does for all other such fields). At this point I've
lost faith and I'm willing to call it an LLVM bug, pending some further
investigation, preferrably with someone who knows LLVM better.

Since this is causing segfaults on some systems, I'm reverting the patch.