Add an initial version of the multicore backend (#204)
Turns out, that having MDImp in place, we can reuse a lot of the
CUDA-related infrastructure for generating multicore CPU code. With this
patch (and some hacks to make MDImp less restrictive) the ray tracer
example takes 98s to generate a 250x250 image using 16 threads, as opposed
to 584s when running with the single-threaded backend.