Differentiate through case expressions with inactive scrutinees
While supporting differentiation for sum/variant types is difficult, it
doesn't mean that we can't differentiate any programs that contain
`case` expressions. If the scrutinee contains no active variables, then
this is equivalent to simply selecting one of the paths through the
program, without any effect on the tangent values.