Avoid loading prelude from the default path
We now bundle the whole source in the compiler binary anyway, so it's
faster to just use that. Additionally, this should allow us to easily
install the `dex` binary without having to carry the `prelude.dx` file
around. The change is not that simple, because it required an update to
the caching mechanism, which now also uses global XDG directories.