Allow explicit arrow annotations for everything except -> and =>
In the spirit of not letting perfect be the enemy of good, this
implements a simpler version of explicit lambda arrows. Arrows
are now allowed to replace the . terminator of lambda expressions,
but -> and => cannot be used there, since . is still used for normal
explicit lambdas and => should use for instead of a lambda expression.

Alternatives we may want to consider in the future:
- Replacing . with -> as an explicit separator. This is more similar
  to Haskell, and is more consistent, but also requires changing a
  lot of existing code.
- Allowing table lambdas, which are currently an implementation
  detail but maybe in the future would be something we would want
  to expose as a lightweight view.
- Inferring arrow type during type inference, so that we can use
  the . separator even for implicit arrows.
- Using braces syntax, such as `\{x} y.`, to denote that x is an
  implicit argument and y is explicit. Right now this is ambiguous
  because it could also be interpreted as unpacking a record
  pattern.