Implement a parallelism flattening pass
This will scan the program for nested parallelism and attempt to convert
it to flat parallelism. For example, the following:
```
for i:(Fin 100).
  x = f i
  for j:(Fin 100000).
    g x i j
```
will get converted into:
```
xs = for i:(Fin 100). fi
for (i,j):(Fin 100 & Fin 100000).
  g xs.i i j
```
This exposes a significant amount of parallelism in the second loop,
which was previously hidden in the short first loop.

Unfortunately such flattening comes with a cost, and the
downsides are clearly visible even in the example above. While the
original program has only required a very small workspace to compute
each output element, the second one materializes each value of a
loop-local variable in an array, potentially leading to significantly
higher memory usage.

Due to that, we should be careful about the loops we choose to flatten,
and we should stop as soon as we decide that we've exposed enough
parallelism at the top level. This commit only implements a very
aggressive flattening scheme, which we'll definitely want to refine at
some point, but the implementation should be flexible enough to make it
easy for us to play around with different choices.

Also note that while this flattening can serve as a way to pull out
allocation of large thread-local intermediates out of each kernel, it
does so in an inefficient way. This is because we always end up scaling
up memory usage by the size of the whole iteration domain, while it
would be sufficient to scale by the maximum number of threads active at
any point. Hence, we'll need to implement allocation lifting anyway.

Finally, there is one more way to reduce the amount of memory needed:
instead of extracting perfect parallelism with no inter-thread
synchronization, we could take advantage of the fact that e.g. on GPUs
blocks of as many as 1024 threads can synchronize without incurring
large penalties. This way, the O(1000)-sized intermediate arrays could
be easily stored in fast local memory.