Make the kernel language more flexible
This exposes a number of important variables in the kernel language,
which expose the concept of a workgroup to the Imp representation. While
this does not come with any significant upsides at the moment, it is a
necessary prerequisite for the efficient parallel reductions that I am
hoping to submit soon (I have most of the implementation ready).

Here's a summary of the changes:
* It is no longer the responsibility of `ILaunch` to actually cover the
  iteration domain with threads. `ILaunch` is supposed to launch as many
  threads as can concurrently execute on the given accelerator, and it is
  the responsibility of each thread to figure out its iteration domain
  (which might involve multiple points of the full domain).
* Kernel threads now expose a number of useful variables such as
  (intra-workgroup) thread id, workgroup id, workgroup size, total number
  of threads.
* Given those constraints, CUDA kernels are now using grid-stride loops,
  while CPU kernels still iterate over the domain in contiguous chunks.
  This required an addition of an Imp-lowering-specific index range type
  called `ParIndexRange` and the corresponding constructor `ParIndexCon`.