Cache CUDA kernels on the device
Previously whenever we wanted to launch a CUDA kernel we would emit LLVM
code to load it onto the device, call into it, and unload it again.
While this was a good way to get started, loading the code involves a
JIT and can take hundreds of milliseconds, i.e. orders of magnitude more
than the actual GPU execution. This made it almost impossible to
benchmark our CUDA performance, because the run-time was heavily
dominated by code processing.

This change makes it so that the references to GPU code are lazily
created, but are cached for as long as the object files remain in
memory. The liftime management is mostly automatic, thanks to the
special `llvm.global_dtors` variable, which registers our destructors in
the ELF files. Unfortunately, the JIT interface we use does not support
that option, so we have to emulate it in there, but it's not too
difficult.

Also, now that we have this mechanism worked out, we could consider
caching GPU memory allocations using a technique similar to the one
[used in Futhark](https://futhark-lang.org/blog/2018-01-28-how-futhark-manages-gpu-memory.html).