Add ProjectElt atom, which extracts specific elements from constructors.
To try to reduce the surface for bugs, ProjectElt syntactically requires its
argument to be a Var, indexed by a nonempty list of indices. This means that
substituting into a ProjectElt requires immediately reducing it.

Note that getting the type of a ProjectElt atom is a bit subtle, because if
we are extracting a value from an existential ADT DataCon, the type of the
projected result may itelf include earlier bindings in the DataCon, which
must also be converted to ProjectElt atoms.