Modify implicit args, reduce projections, get list length example working.
Modify implicit args: To allow function calls inside type annotations, this
change makes it so that any lowercase name that is used in function position
of an application (e.g. `f x`) will NOT be added as an automatic implicit type
parameter. This is an improvement because such a type application would never
typecheck anyway if `f` was inferred to have type `Type`. This also matches
the behavior of Idris.

Reduce projections: We assume during typechecking that `getType` always returns
a fully reduced type. But `getType` of a projection may produce another
projection with the same root variable. Thus, whenver we create a new
projection, we have to reduce the variable. (It's not clear that this is the
best way to do this, but it seems to work for now.)

Get list extraction example working: With these two changes, it becomes possible
to construct functions that do dependent projections, for instance by converting
a `List a` into a table. The syntax is a bit unwieldy for this, but that should
be easy to fix.