Create a self-contained abstraction for LLVM JIT infrastructure
Previously, our JIT code was heavily specialized to one-shot
evaluation which is fine when all we care about is interpretation.
But one interesting use case I would like us to support is surfacing
JIT-compiled reentrant (in the sense that they can be called multiple
times) kernels in Python, which will require us to keep the compiled
code around for a while longer.

The new abstractions all live in `LLVM.JIT`. First, there's the `JIT`
ADT which represents an instance of the LLVM JIT objects. There's not
much you can do with that one except for creating and destroying it.
However, its real purpose is so that one can use it to compile and
register `NativeModule`s (potentially multiple at the same time). Such
compiled modules can be queried for function pointers and won't be
unloaded until explicitly asked to do so.

Note that this is also a useful step for compilation caching. While the
current implementation does not support linking accross modules, it
would be a straightforward extension to the symbol resolvers.