Make it possible to export JITed Dex functions to Python
The API is still far from ergonomic, because it returns raw function
pointers, with signatures derived from pretty obscure Haskell code. The
next step will be to clean up the expected signatures and handle all the
destination allocation and dereferencing internally.

This is a big-ish change, mostly because it required a further refactor
of our compilation pipeline, and a few additions to the llvm-hs APIs
(non-bracketed context and module management). Until those changes are
merged upstream, I changed `stack.yaml` to point to the PR branch in my
fork.