Make a "virtual stdout" stream that we can write to using the IO effect.
We introduce a special global variable, `OUT_STREAM_PTR`, which at runtime
resolves to a C stdio `FILE*` pointer. The runtime creates a fresh pipe every
time it runs an LLVM program. It puts the write end in `OUT_STREAM_PTR`, and
captures what comes out the other end in a separate thread while the LLVM
program runs.

The Haskell thread that reads the output needs to be able to concurrently with
the LLVM program in case the pipe fills up. To ensure this can happen, we use
the `-threaded` GHC option.